+++ mobile/app/wallet/assets/ethereum/config/ethereum-config.ts
@@ -5,9 +5,11 @@ const isTestNet = Network === "TEST";
 interface Config {
   IsTestNet: boolean;
   coinTypeIndex: string;
+  chain: string;
 }
 
 export const config: Config = {
   IsTestNet: isTestNet,
   coinTypeIndex: isTestNet ? "1" : "60",
+  chain: isTestNet ? "goerli" : "mainnet",
 };
