+++ mobile/package.json
@@ -22,7 +22,6 @@
     "elliptic": "^6.5.4",
     "ethereumjs-tx": "^2.1.2",
     "ethereumjs-util": "^7.1.5",
-    "ethers": "^5.6.9",
     "events": "^1.1.1",
     "expo": "^45.0.0",
     "expo-local-authentication": "~12.2.0",
