+++ mobile/packages/blockchain-api-client/src/provider/alchemy/http.ts
@@ -1,7 +1,6 @@
 import { fetchFrom } from '../../base/http';
 
 export const fetchFromAlchemy = async <T>(url: string, method: Method, params?: any[]): Promise<T> => {
-  console.log(url + apiKey, { body: { id: 1, jsonrpc: '2.0', params, method } });
   return fetchFrom(url + apiKey, { body: { id: 1, jsonrpc: '2.0', params, method } });
 };
